reorganize ProductProfile component files
