+++ client/src/components/ProductProfile/TechItem.jsx
