+++ client/src/components/ProductProfile/TechList.jsx
